requirements.txt
asgiref==3.7.2
configparser==6.0.0
Django==5.0.2
mysql-connector==2.2.9
mysql-connector-python==8.3.0
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
